bcm2708: Add rpi-pwm driver to machine definition
This patch is in preparation for adding a misc driver for the Raspberry Pi
PWM pin.  It lets the kernel know that this platform supports a device
called "bcm2708_pcm", so that if such a driver is encountered it can be
automatically loaded.

The driver requires no resources as all offsets are hardcoded within the
driver itself.

Conflicts:

	arch/arm/mach-bcm2708/bcm2708.c